Makes sure that session is destroyed when expired
Also tidyed up strurture of session and logout functions.
