+++ middleware.ts
@@ -4,7 +4,9 @@ import type { NextFetchEvent, NextRequest } from "next/server"
 import loadUserToken from "./app/auth/callback/adgangsplatformen/loadUserToken"
 import goConfig from "./lib/config/goConfig"
 import {
+  accessTokenIsExpired,
   accessTokenShouldBeRefreshed,
+  destroySession,
   getDplCmsSessionCookie,
   getSession,
   saveAdgangsplatformenSession,
@@ -20,14 +22,10 @@ export async function middleware(request: NextRequest, event: NextFetchEvent) {
 
   const session = await getSession({ request, response })
 
-  // TODO: Check if we need to destroy the session if the refresh token is expired.
-  // const session = await getIronSession<TSessionData>(request, response, sessionOptions);
-  // const isExpired = accessTokenIsExpired(session);
-  // console.log({ isExpired });
-  // if (isExpired) {
-  //   session.destroy();
-  //   return NextResponse.redirect(new URL('/', request.url), { headers: response.headers });
-  // }
+  // If the session is expired then destroy it.
+  if (accessTokenIsExpired(session)) {
+    destroySession(session)
+  }
 
   // If the session is not logged in we will try to see
   // if we have an ongoing Adgangsplatformen Drupal session.
@@ -43,7 +41,7 @@ export async function middleware(request: NextRequest, event: NextFetchEvent) {
   if (session.isLoggedIn && session.type === "adgangsplatformen") {
     const sessionCookie = await getDplCmsSessionCookie()
     if (!sessionCookie) {
-      session.destroy()
+      destroySession(session)
     }
   }
